add GD, ISTA, FISTA as children of PGA
